+++ components/Menu/index.tsx
@@ -64,10 +64,13 @@ const Menu: React.FC = () => {
   };
 
   return (
-    <div className="p-2 select-none hidden md:block md:w-[17.5rem]">
+    <div className="p-2 select-none hidden md:block md:w-[17.5rem] transition-colors dark:bg-zinc-900">
       <div
         onClick={onAddChannel}
-        className="rounded-lg cursor-pointer flex bg-[#678fff] h-12 text-white mb-2 transition-colors justify-center items-center hover:bg-[#678fff]/80"
+        className={classNames(
+          "rounded-lg text-white cursor-pointer flex h-12 mb-2 transition-all justify-center items-center",
+          "bg-gradient-to-r from-cyan-500 to-blue-500 bg-magic-size hover:bg-magic-position"
+        )}
       >
         {t("new-chat")}
       </div>
@@ -77,21 +80,32 @@ const Menu: React.FC = () => {
             key={item.channel_id}
             onClick={() => onChangeChannel(item.channel_id)}
             className={classNames(
-              "rounded-lg cursor-pointer mb-1 overflow-hidden relative flex flex-col h-16 text-xs text-base-color px-[0.5rem] transition-colors gap-1 group justify-center hover:bg-menu-hover",
-              { "!bg-menu-active": item.channel_id === channel.activeId }
+              "rounded-lg cursor-pointer mb-1 overflow-hidden relative flex flex-col h-16 text-xs text-base-color px-[0.5rem] transition-colors gap-1 group justify-center",
+              "hover:bg-menu-hover dark:hover:bg-neutral-800",
+              {
+                "!bg-menu-active dark:!bg-slate-700":
+                  item.channel_id === channel.activeId,
+              }
             )}
           >
             <div className="flex justify-between items-center">
-              <div className="font-medium text-sm text-ellipsis max-w-[26ch] pl-5 relative overflow-hidden whitespace-nowrap">
+              <div className="font-medium text-sm text-ellipsis max-w-[26ch] pl-5 relative overflow-hidden whitespace-nowrap transition-colors dark:text-white">
                 <BsChatSquareText className="top-[50%] left-0 translate-y-[-50%] absolute" />
                 {item.channel_name || t("new-conversation")}
               </div>
             </div>
-            <div className="flex text-[#858b96] justify-between group-hover:text-[#6e737b]">
+            <div
+              className={classNames(
+                "flex text-[#858b96] justify-between transition-all group-hover:text-[#6e737b] dark:text-neutral-500 dark:group-hover:text-neutral-400",
+                {
+                  "dark:text-neutral-400": item.channel_id === channel.activeId,
+                }
+              )}
+            >
               <div>
                 {item.chat_list.length} {t("messages")}
               </div>
-              <div className="transition-opacity text-[#858b96] group-hover:opacity-0">
+              <div className="group-hover:opacity-0">
                 {item.chat_list.length
                   ? item.chat_list.at(-1)?.time
                     ? format(
@@ -108,7 +122,7 @@ const Menu: React.FC = () => {
               trigger={
                 <div
                   onClick={stopPropagation}
-                  className="opacity-0 transition-all right-[-2rem] absolute group-hover:opacity-100 group-hover:right-2"
+                  className="opacity-0 transition-all right-[-2rem] absolute dark:text-white group-hover:opacity-100 group-hover:right-2"
                 >
                   <AiOutlineDelete size={20} />
                 </div>
@@ -118,12 +132,12 @@ const Menu: React.FC = () => {
           </div>
         ))}
       </div>
-      <div className="h-[6.5rem] flex flex-col gap-2 border-t pt-2">
+      <div className="h-[6.5rem] flex flex-col gap-2 border-t dark:border-neutral-600 pt-2">
         <Confirm
           title={t("clear-all-conversation")}
           content={t("clear-conversation")}
           trigger={
-            <div className="h-11 rounded-md transition-colors text-sm hover:bg-menu-hover cursor-pointer flex items-center gap-2 px-2">
+            <div className="h-11 rounded-md transition-colors text-sm hover:bg-menu-hover dark:hover:bg-neutral-800 cursor-pointer flex items-center gap-2 px-2">
               <AiOutlineDelete size={16} /> {t("clear-all-conversation")}
             </div>
           }
@@ -132,7 +146,7 @@ const Menu: React.FC = () => {
         <a
           href="https://github.com/Peek-A-Booo/L-GPT"
           target="_blank"
-          className="h-11 rounded-md transition-colors text-sm hover:bg-menu-hover cursor-pointer flex items-center gap-2 px-2"
+          className="h-11 rounded-md transition-colors text-sm hover:bg-menu-hover dark:hover:bg-neutral-800 cursor-pointer flex items-center gap-2 px-2"
         >
           <AiFillGithub size={16} /> Github
         </a>