+++ components/Navbar/setting.tsx
@@ -1,9 +1,33 @@
 import * as React from "react";
 import { useTranslation } from "next-i18next";
-import { Modal } from "@/components";
+import { useTheme } from "next-themes";
+import { Modal, Select } from "@/components";
 import { useOpenAIKey, useProxy } from "@/hooks";
 
+const moduleOptions = [
+  {
+    label: "gpt-3.5",
+    value: "gpt-3.5",
+  },
+];
+
+const themeOptions = [
+  {
+    label: "light",
+    value: "light",
+  },
+  {
+    label: "dark",
+    value: "dark",
+  },
+  {
+    label: "auto",
+    value: "auto",
+  },
+];
+
 const Setting = React.forwardRef((_, forwardedRef) => {
+  const { theme, setTheme } = useTheme();
   const [open, setOpen] = React.useState(false);
   const [openAIKey, setOpenAIKey] = useOpenAIKey();
   const [proxyUrl, setProxyUrl] = useProxy();
@@ -15,6 +39,7 @@ const Setting = React.forwardRef((_, forwardedRef) => {
   React.useImperativeHandle(forwardedRef, () => ({
     init() {
       setOpen(true);
+      console.log(theme, "theme");
     },
   }));
 
@@ -27,10 +52,10 @@ const Setting = React.forwardRef((_, forwardedRef) => {
       onClose={onClose}
     >
       <div className="border-b flex border-slate-100 py-2 px-1 items-center justify-between">
-        <div className="font-semibold text-sm">API key</div>
+        <div className="font-semibold text-sm text-black">API key</div>
         <div className="text-xs">
           <input
-            className="border rounded-md p-2"
+            className="border rounded-md p-2 dark:bg-white"
             type="password"
             placeholder={t("set-openai-key") as string}
             value={openAIKey}
@@ -39,10 +64,10 @@ const Setting = React.forwardRef((_, forwardedRef) => {
         </div>
       </div>
       <div className="border-b flex border-slate-100 py-2 px-1 items-center justify-between">
-        <div className="font-semibold text-sm">{t("proxy-url")}</div>
+        <div className="font-semibold text-sm text-black">{t("proxy-url")}</div>
         <div className="text-xs">
           <input
-            className="border rounded-md p-2"
+            className="border rounded-md p-2 dark:bg-white"
             placeholder={t("set-proxy-url") as string}
             value={proxyUrl}
             onChange={(e) => setProxyUrl(e.target.value)}
@@ -50,19 +75,20 @@ const Setting = React.forwardRef((_, forwardedRef) => {
         </div>
       </div>
       <div className="border-b flex border-slate-100 py-2 px-1 items-center justify-between">
-        <div className="font-semibold text-sm">{t("model")}</div>
+        <div className="font-semibold text-sm text-black">{t("model")}</div>
         <div className="text-xs">
-          <select className="border rounded-md p-2">
-            <option value="gpt-3.5">gpt-3.5</option>
-          </select>
+          <Select options={moduleOptions} />
         </div>
       </div>
       <div className="border-b flex border-slate-100 py-2 px-1 items-center justify-between">
-        <div className="font-semibold text-sm">{t("theme")}</div>
+        <div className="font-semibold text-sm text-black">{t("theme")}</div>
         <div className="text-xs">
-          <select className="border rounded-md p-2">
-            <option value="auto">auto</option>
-          </select>
+          <Select
+            options={themeOptions}
+            placeholder="Select"
+            value={theme}
+            onChange={setTheme}
+          />
         </div>
       </div>
     </Modal>