+++ components/Select/index.tsx
@@ -1,0 +1,82 @@
+import * as React from "react";
+import * as Select from "@radix-ui/react-select";
+import classnames from "classnames";
+import { AiFillCaretDown } from "react-icons/ai";
+
+interface SelectItemProps extends React.PropsWithChildren {
+  className?: string | undefined;
+  value: any;
+  disabled?: boolean;
+}
+
+type Options = {
+  label: string;
+  value: any;
+};
+
+export interface SelectProps {
+  onChange?: (value: any) => void;
+  options: Options[];
+  placeholder?: string;
+  value?: any;
+}
+
+const LSelect: React.FC<SelectProps> = ({
+  onChange,
+  options,
+  placeholder,
+  value,
+}) => {
+  return (
+    // className="inline-flex items-center justify-center rounded px-[15px] text-[13px] leading-none h-[35px] gap-[5px] bg-white text-violet11 shadow-[0_2px_10px] shadow-black/10 hover:bg-mauve3 focus:shadow-[0_0_0_2px] focus:shadow-black data-[placeholder]:text-violet9 outline-none"
+    <Select.Root value={value} onValueChange={onChange}>
+      <Select.Trigger className="inline-flex items-center justify-center rounded px-[15px] text-[13px] leading-none h-[35px] gap-[5px] bg-white text-violet11 shadow-[0_2px_10px] shadow-black/10 hover:bg-mauve3 focus:shadow-[0_0_0_2px] focus:shadow-black data-[placeholder]:text-violet9 outline-none">
+        <Select.Value placeholder={placeholder} />
+        <Select.Icon className="text-violet11">
+          <AiFillCaretDown />
+        </Select.Icon>
+      </Select.Trigger>
+      <Select.Portal>
+        <Select.Content className="overflow-hidden z-[1999] bg-white rounded-md shadow-[0px_10px_38px_-10px_rgba(22,_23,_24,_0.35),0px_10px_20px_-15px_rgba(22,_23,_24,_0.2)]">
+          <Select.ScrollUpButton className="flex items-center justify-center h-[25px] bg-white text-violet11 cursor-default">
+            up
+          </Select.ScrollUpButton>
+          <Select.Viewport className="p-[5px]">
+            {options.map((item) => (
+              <SelectItem key={item.value} value={item.value}>
+                {item.label}
+              </SelectItem>
+            ))}
+          </Select.Viewport>
+          <Select.ScrollDownButton className="flex items-center justify-center h-[25px] bg-white text-violet11 cursor-default">
+            down
+          </Select.ScrollDownButton>
+        </Select.Content>
+      </Select.Portal>
+    </Select.Root>
+  );
+};
+
+const SelectItem = React.forwardRef<any, SelectItemProps>(
+  ({ children, className, ...props }, forwardedRef) => {
+    return (
+      <Select.Item
+        className={classnames(
+          "text-[13px] leading-none text-violet11 rounded-[3px] flex items-center h-[25px] pr-[35px] pl-[25px] relative select-none data-[disabled]:text-mauve8 data-[disabled]:pointer-events-none data-[highlighted]:outline-none data-[highlighted]:bg-violet9 data-[highlighted]:text-violet1",
+          className
+        )}
+        {...props}
+        ref={forwardedRef}
+      >
+        <Select.ItemText>{children}</Select.ItemText>
+        <Select.ItemIndicator className="absolute left-0 w-[25px] inline-flex items-center justify-center">
+          {/* <CheckIcon /> */}
+        </Select.ItemIndicator>
+      </Select.Item>
+    );
+  }
+);
+
+SelectItem.displayName = "SelectItem";
+
+export default LSelect;