+++ components/index.ts
@@ -12,5 +12,6 @@ export {
 } from "./ScrollToBottoms";
 export { default as Input } from "./Input";
 export { default as Button } from "./Button";
-export { default as Confirm } from "./Confirm";
 export { default as Modal } from "./Modal";
+export { default as Confirm } from "./Modal/confirm";
+export { default as Select } from "./Select";